Changing behavior using IPFS
We are separating the environment we use to upload prod or
staging releases. For staging, we are going to use a different
pfs server. Therefore Workflow's that need staging Jobs will
only be triggered if the branch is not main.
For prod we keep using the same approach regarding IPFS, with the
difference that Job will only be triggered with the branch "main"